[docs] [train] copy edits to E2E ingest example (#40176)
Trying to add copy edits as a commit. @stephanie-wang, sorry if this
didn't work!

The code examples still need to be updated to the new Train GA APIs.
References to AIR are now out of date.

## Why are these changes needed?

Copy edits to make the style more consistent with the style guide we
follow.

## Related issue number

https://github.com/ray-project/ray/pull/38082

## Checks

- [x] I've signed off every commit(by using the -s flag, i.e., `git
commit -s`) in this PR.
- [ ] I've run `scripts/format.sh` to lint the changes in this PR.
- [ ] I've included any doc changes needed for
https://docs.ray.io/en/master/.
- [ ] I've added any new APIs to the API Reference. For example, if I
added a
method in Tune, I've added it in `doc/source/tune/api/` under the
           corresponding `.rst` file.
- [ ] I've made sure the tests are passing. Note that there might be a
few flaky tests, see the recent failures at https://flakey-tests.ray.io/
- Testing Strategy
   - [ ] Unit tests
   - [ ] Release tests
   - [ ] This PR is not tested :(

---------

Signed-off-by: angelinalg <122562471+angelinalg@users.noreply.github.com>